refactor: conmponents 폴더 정리
