+++ client/src/components/Room/Message.tsx
@@ -1,4 +1,4 @@
-import { ChatEvent, MessageInterface } from '../types/Message';
+import { ChatEvent, MessageInterface } from '../../types/Message';
 
 // This just needs to be here so that these colors get bundled in the final distribution.
 // The userColor is actually assigned on the server.
