+++ client/src/components/Room/RoomInfo.tsx
@@ -1,4 +1,4 @@
-import ExitButton from './buttons/ExitButton';
+import ExitButton from '../buttons/ExitButton';
 import { FaRegCopy } from 'react-icons/fa6';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -15,7 +15,7 @@ export default function RoomInfo({ roomCode }: { roomCode: string }) {
   return (
     <>
       <div className="flex w-full items-center justify-between">
-        <div className="text-text_default flex flex-row items-center rounded-lg">
+        <div className="flex flex-row items-center rounded-lg text-text_default">
           <div className="pr-3 text-xl font-bold">{roomCode}</div>
           <button onClick={copyToClipboard}>
             <FaRegCopy />
