+++ client/src/components/Room/ScoreBoardModal/Result.tsx
@@ -1,6 +1,6 @@
 import { RxCheck, RxCross1, RxBorderSolid } from 'react-icons/rx';
 
-import { ScoreResult } from '../../types/ScoreType';
+import { ScoreResult } from '../../../types/ScoreType';
 
 interface ResultProps {
   result: ScoreResult;
@@ -8,7 +8,7 @@ interface ResultProps {
 
 export default function Result({ result }: ResultProps) {
   return (
-    <div className="flex flex-1 flex-row justify-around items-center">
+    <div className="flex flex-1 flex-row items-center justify-around">
       {result === ScoreResult.CORRECT ? (
         <RxCheck className="text-green" strokeWidth={2} size={20} />
       ) : result === ScoreResult.WRONG ? (
