+++ client/src/components/Room/ScoreBoardModal/ScoreBoard.tsx
@@ -1,4 +1,4 @@
-import { Score } from '../../types/ScoreType';
+import { Score } from '../../../types/ScoreType';
 import Players from './Players';
 
 interface ScoreBoardProps {
@@ -7,7 +7,7 @@ interface ScoreBoardProps {
 
 export default function ScoreBoard({ scores }: ScoreBoardProps) {
   return (
-    <ul className="text-text_default my-5 flex w-full flex-col overflow-auto text-sm font-medium">
+    <ul className="my-5 flex w-full flex-col overflow-auto text-sm font-medium text-text_default">
       {scores.players.map((playerScore, index) => (
         <Players key={index} playerScore={playerScore} />
       ))}
